restructure the profile avatar so the area around the avatar is clickable
