+++ src/pages/home/sidebar/AvatarWithOptionalStatus.tsx
@@ -1,11 +1,8 @@
 import React from 'react';
 import {View} from 'react-native';
-import PressableWithoutFeedback from '@components/Pressable/PressableWithoutFeedback';
 import Text from '@components/Text';
-import useLocalize from '@hooks/useLocalize';
 import useThemeStyles from '@hooks/useThemeStyles';
-import CONST from '@src/CONST';
-import PressableAvatarWithIndicator from './PressableAvatarWithIndicator';
+import ProfileAvatarWithIndicator from './ProfileAvatarWithIndicator';
 
 type AvatarWithOptionalStatusProps = {
     /** Emoji status */
@@ -13,34 +10,22 @@ type AvatarWithOptionalStatusProps = {
 
     /** Whether the avatar is selected */
     isSelected?: boolean;
-
-    /** Callback called when the avatar or status icon is pressed */
-    onPress?: () => void;
 };
 
-function AvatarWithOptionalStatus({emojiStatus = '', isSelected = false, onPress}: AvatarWithOptionalStatusProps) {
+function AvatarWithOptionalStatus({emojiStatus = '', isSelected = false}: AvatarWithOptionalStatusProps) {
     const styles = useThemeStyles();
-    const {translate} = useLocalize();
 
     return (
         <View style={styles.sidebarStatusAvatarContainer}>
-            <PressableAvatarWithIndicator
-                isSelected={isSelected}
-                onPress={onPress}
-            />
-            <PressableWithoutFeedback
-                accessibilityLabel={translate('sidebarScreen.buttonMySettings')}
-                role={CONST.ROLE.BUTTON}
-                onPress={onPress}
-                style={[styles.sidebarStatusAvatar]}
-            >
+            <ProfileAvatarWithIndicator isSelected={isSelected} />
+            <View style={[styles.sidebarStatusAvatar]}>
                 <Text
                     style={styles.emojiStatusLHN}
                     numberOfLines={1}
                 >
                     {emojiStatus}
                 </Text>
-            </PressableWithoutFeedback>
+            </View>
         </View>
     );
 }
